Fixed duplicate iters
