+++ src/main/scala/dpla/ingestion3/harvesters/file/NaraFileHarvester.scala
@@ -28,26 +28,6 @@ class NaraFileHarvester(
     conf: i3Conf
 ) extends LocalHarvester(spark, shortName, conf) {
 
-  /** Case class hold the parsed value from a given FileResult
-    */
-  case class ParsedResult(id: String, item: String)
-
-  /** Case class to hold the results of a file
-    *
-    * @param entryName
-    *   Path of the entry in the file
-    * @param data
-    *   Holds the data for the entry, or None if it's a directory.
-    * @param bufferedData
-    *   Holds a buffered reader for the entry if it's too large to be held in
-    *   memory.
-    */
-  case class FileResult(
-      entryName: String,
-      data: Option[Array[Byte]],
-      bufferedData: Option[BufferedReader] = None
-  )
-
   lazy val naraSchema: Schema =
     new Schema.Parser()
       .parse(new FlatFileIO().readFileAsString("/avro/NaraOriginalRecord.avsc"))
@@ -214,9 +194,9 @@ class NaraFileHarvester(
 
     logger.info(s"Writing harvest tmp output to $naraTmp")
 
-    // FIXME This assumes files on local file system and not on S3. Files should be able to read off of S3.
+
     val inFile = new File(conf.harvest.endpoint.getOrElse("in"))
-    // FIXME Deletes are tracked in files alongside harvest data, this should be done in a more sustainable way
+
     val deletes = new File(inFile, "/deletes/")
 
     if (inFile.isDirectory)
@@ -310,19 +290,16 @@ class NaraFileHarvester(
         )
       )
 
-    val recordCount = iter(inputStream).map(tarResult =>
+    FileHarvester.iter(inputStream).foreach(tarResult =>
       handleFile(tarResult, unixEpoch, file.getName) match {
         case Failure(exception) =>
           val logger = LogManager.getLogger(this.getClass)
           logger
             .error(s"Caught exception on ${tarResult.entryName}.", exception)
-          0
-        case Success(count) =>
-          count
-      }).sum
 
-    val logger = LogManager.getLogger(this.getClass)
-    logger.info(s"Harvested $recordCount records from ${file.getName}")
+        case Success(count) => ()
+      }
+    )
 
     IOUtils.closeQuietly(inputStream)
   }
