+++ src/main/scala/dpla/ingestion3/harvesters/file/VaFileHarvester.scala
@@ -37,9 +37,6 @@ class VaFileHarvester(
     * @return
     *   ZipInputstream of the zip contents
     *
-    * TODO: Because we're only handling zips in this class, and they should
-    * already be filtered by the FilenameFilter, I wonder if we even need the
-    * match statement here.
     */
   def getInputStream(file: File): Option[ZipInputStream] =
     file.getName match {
@@ -76,28 +73,6 @@ class VaFileHarvester(
         }
     }
 
-  /** Implements a stream of files from the zip Can't use @tailrec here because
-    * the compiler can't recognize it as tail recursive, but this won't blow the
-    * stack.
-    *
-    * @param zipInputStream
-    * @return
-    *   Lazy stream of zip records
-    */
-  def iter(zipInputStream: ZipInputStream): LazyList[FileResult] =
-    Option(zipInputStream.getNextEntry) match {
-      case None =>
-        LazyList.empty
-      case Some(entry) =>
-        val result =
-          if (entry.isDirectory || !entry.getName.endsWith(".xml"))
-            None
-          else {
-            Some(IOUtils.toByteArray(zipInputStream, entry.getSize))
-          }
-        FileResult(entry.getName, result) #:: iter(zipInputStream)
-    }
-
   /** Executes the Digital Virginias harvest
     */
   override def localHarvest(): DataFrame = {
@@ -112,7 +87,7 @@ class VaFileHarvester(
           .getOrElse(
             throw new IllegalArgumentException("Couldn't load ZIP files.")
           )
-        iter(inputStream).foreach(result => {
+        FileHarvester.iter(inputStream).foreach(result => {
           handleFile(result, unixEpoch) match {
             case Failure(exception) =>
               LogManager
